Make 1 file for firebase functions
