+++ src/components/signup-flow/input/Input.tsx
@@ -1,9 +1,7 @@
 import React from "react";
 import { Field } from "formik";
-import * as Yup from "yup";
-import { sendPasswordResetEmail } from "firebase/auth";
 
-import { auth } from "@/pages/api/firebase";
+import { handleForgotPassword } from "src/lib/firebaseFunctions";
 
 import stl from "./Input.module.scss";
 
@@ -15,25 +13,6 @@ interface Props {
 }
 
 const Input = ({ formikProps, id, placeholder, method }: Props) => {
-  const emailSchema = Yup.object().shape({
-    email: Yup.string()
-      .email("Invalid email address")
-      .required("Email is required"),
-  });
-
-  const handleForgotPassword = async () => {
-    const email = formikProps.values.email;
-    try {
-      await emailSchema.validate({ email });
-      sendPasswordResetEmail(auth, email)
-        .then(() => console.log("Password Reset link sent to", email))
-        .catch((err) =>
-          console.log("Error while sending Password Reset link", err)
-        );
-    } catch (error) {
-      console.log("Error from Validation:", error);
-    }
-  };
   const handleFocus = (e: any) => {
     const ele = e.target;
     ele.classList.add(stl.focused);
@@ -59,7 +38,10 @@ const Input = ({ formikProps, id, placeholder, method }: Props) => {
         {formikProps.touched[id] && formikProps.errors[id]}
       </span>
       {method === "signin" && id === "password" && (
-        <span onClick={handleForgotPassword} className={stl.forgotPassword}>
+        <span
+          onClick={() => handleForgotPassword(formikProps)}
+          className={stl.forgotPassword}
+        >
           Forgot Password?
         </span>
       )}
