+++ src/components/verify-msg/VerifyMsg.module.scss
@@ -10,4 +10,24 @@
   span {
     text-decoration: underline;
   }
+
+  .btn {
+    border: 1px solid #1e90ff;
+    border-radius: 7px;
+    background: none;
+    padding: 10px;
+    font-size: 1.2rem;
+    cursor: pointer;
+    transition: all 0.3s;
+
+    &:hover {
+      background: #1e90ff;
+      border-color: transparent;
+      color: #fff;
+    }
+
+    &:active {
+      transform: scale(0.9);
+    }
+  }
 }
