+++ src/components/verify-msg/VerifyMsg.tsx
@@ -1,5 +1,7 @@
 import React from "react";
 
+import { sendVerificationEmail } from "src/lib/firebaseFunctions";
+
 import stl from "./VerifyMsg.module.scss";
 
 const VerifyMsg = ({ email }: any) => {
@@ -9,6 +11,9 @@ const VerifyMsg = ({ email }: any) => {
         Verification email sent to <span>{email}</span>.
       </h2>
       <h2>Verify your Email to Continue!</h2>
+      <button onClick={sendVerificationEmail} className={stl.btn}>
+        Resend
+      </button>
     </div>
   );
 };
