+++ tsconfig.json
@@ -26,7 +26,7 @@
     "**/*.ts",
     "**/*.tsx",
     "pages/api/firebase.js",
-    "src/useClickOutside.js"
+    "src/lib/useClickOutside.js"
   ],
   "exclude": ["node_modules"]
 }
